+++ spec/integration/preview/sync/service/document_spec.rb
@@ -210,10 +210,10 @@ describe 'Document' do
 
     expect {
       @client.preview.sync.services('ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
-                          .documents('ETaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa').update(data: {})
+                          .documents('ETaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa').update(data: JSON.parse('{}'))
     }.to raise_exception(Twilio::REST::TwilioError)
 
-    values = {'Data' => Twilio.serialize_object({}), }
+    values = {'Data' => Twilio.serialize_object(JSON.parse('{}')), }
     expect(
     @holodeck.has_request?(Holodeck::Request.new(
         method: 'post',
@@ -245,7 +245,7 @@ describe 'Document' do
     ))
 
     actual = @client.preview.sync.services('ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
-                                 .documents('ETaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa').update(data: {})
+                                 .documents('ETaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa').update(data: JSON.parse('{}'))
 
     expect(actual).to_not eq(nil)
   end