+++ spec/integration/preview/sync/service/sync_map/sync_map_item_spec.rb
@@ -86,10 +86,10 @@ describe 'SyncMapItem' do
     expect {
       @client.preview.sync.services('ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
                           .sync_maps('MPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
-                          .sync_map_items.create(key: 'key', data: {})
+                          .sync_map_items.create(key: 'key', data: JSON.parse('{}'))
     }.to raise_exception(Twilio::REST::TwilioError)
 
-    values = {'Key' => 'key', 'Data' => Twilio.serialize_object({}), }
+    values = {'Key' => 'key', 'Data' => Twilio.serialize_object(JSON.parse('{}')), }
     expect(
     @holodeck.has_request?(Holodeck::Request.new(
         method: 'post',
@@ -119,7 +119,7 @@ describe 'SyncMapItem' do
 
     actual = @client.preview.sync.services('ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
                                  .sync_maps('MPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
-                                 .sync_map_items.create(key: 'key', data: {})
+                                 .sync_map_items.create(key: 'key', data: JSON.parse('{}'))
 
     expect(actual).to_not eq(nil)
   end
@@ -212,10 +212,10 @@ describe 'SyncMapItem' do
     expect {
       @client.preview.sync.services('ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
                           .sync_maps('MPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
-                          .sync_map_items('key').update(data: {})
+                          .sync_map_items('key').update(data: JSON.parse('{}'))
     }.to raise_exception(Twilio::REST::TwilioError)
 
-    values = {'Data' => Twilio.serialize_object({}), }
+    values = {'Data' => Twilio.serialize_object(JSON.parse('{}')), }
     expect(
     @holodeck.has_request?(Holodeck::Request.new(
         method: 'post',
@@ -245,7 +245,7 @@ describe 'SyncMapItem' do
 
     actual = @client.preview.sync.services('ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
                                  .sync_maps('MPaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
-                                 .sync_map_items('key').update(data: {})
+                                 .sync_map_items('key').update(data: JSON.parse('{}'))
 
     expect(actual).to_not eq(nil)
   end