+++ spec/integration/preview/understand/service/model_build_spec.rb
@@ -35,7 +35,9 @@ describe 'ModelBuild' do
           "service_sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "date_created": "2015-07-30T20:00:00Z",
           "sid": "UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-          "unique_name": "unique_name"
+          "unique_name": "unique_name",
+          "build_duration": null,
+          "error_code": null
       }
       ]
     ))
@@ -106,11 +108,13 @@ describe 'ModelBuild' do
                   "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                   "date_updated": "2015-07-30T20:00:00Z",
                   "url": "https://preview.twilio.com/understand/Services/UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/ModelBuilds/UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                  "status": "enqueued",
+                  "status": "failed",
                   "service_sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                   "date_created": "2015-07-30T20:00:00Z",
                   "sid": "UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                  "unique_name": "unique_name"
+                  "unique_name": "unique_name",
+                  "build_duration": null,
+                  "error_code": 23001
               }
           ]
       }
@@ -151,7 +155,9 @@ describe 'ModelBuild' do
           "service_sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "date_created": "2015-07-30T20:00:00Z",
           "sid": "UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-          "unique_name": "unique_name"
+          "unique_name": "unique_name",
+          "build_duration": null,
+          "error_code": null
       }
       ]
     ))
@@ -186,11 +192,13 @@ describe 'ModelBuild' do
           "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "date_updated": "2015-07-30T20:00:00Z",
           "url": "https://preview.twilio.com/understand/Services/UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/ModelBuilds/UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-          "status": "enqueued",
+          "status": "completed",
           "service_sid": "UAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "date_created": "2015-07-30T20:00:00Z",
           "sid": "UGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-          "unique_name": "unique_name"
+          "unique_name": "unique_name",
+          "build_duration": 100,
+          "error_code": null
       }
       ]
     ))