+++ spec/integration/sync/v1/service/sync_list/sync_list_item_spec.rb
@@ -87,10 +87,10 @@ describe 'SyncListItem' do
     expect {
       @client.sync.v1.services('ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
                      .sync_lists('ESaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
-                     .sync_list_items.create(data: {})
+                     .sync_list_items.create(data: JSON.parse('{}'))
     }.to raise_exception(Twilio::REST::TwilioError)
 
-    values = {'Data' => Twilio.serialize_object({}), }
+    values = {'Data' => Twilio.serialize_object(JSON.parse('{}')), }
     expect(
     @holodeck.has_request?(Holodeck::Request.new(
         method: 'post',
@@ -121,7 +121,7 @@ describe 'SyncListItem' do
 
     actual = @client.sync.v1.services('ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
                             .sync_lists('ESaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa') \
-                            .sync_list_items.create(data: {})
+                            .sync_list_items.create(data: JSON.parse('{}'))
 
     expect(actual).to_not eq(nil)
   end